Compiling OK Pile of changes done from Java to Python

+++ java/src/main/java/com/powsybl/python/commons/PyPowsyblApiHeader.java
@@ -272,7 +272,7 @@ public final class PyPowsyblApiHeader {
     public interface LoadFlowValidationParametersPointer extends PointerBase {
 
         @CField("threshold")
-        int getThreshold();
+        double getThreshold();
 
         @CField("threshold")
         void setThreshold(double threshold);
